finish research work (except for mechs...) rework end turn to calculate actual upkeep so results of randomness are paid for appropriately fix repair so that multiple sources stack correctly bugfixes to get working again

+++ WinFormsApp1/ClassLibrary1/Consts.cs
@@ -74,7 +74,7 @@ namespace ClassLibrary1
         public const double AutoRepair = .65;
         public const double AutoRepairPct = .0169;
 
-        public static double GetDamagedValue(Piece piece, double value, int min)
+        public static double GetDamagedValue(Piece piece, double value, double min)
         {
             if (piece.HasBehavior<IKillable>(out IKillable killable))
                 return min + (value - min) * Math.Pow(killable.HitsCur / killable.HitsMax, 1 - killable.Resilience);
